fix(timeline): remove playback judder from scroll quantisation and sampling phase
The timeline moved unevenly during preview playback while frame delivery was
perfectly healthy: a 139s capture showed 58-61 timeline presents every second
with zero dropped seconds, yet 15.9% of on-time frames were drawn at the wrong
position. Two independent causes, both positional rather than throughput.

1. Sampling was not phase-locked. Playback was sampled from a free-running
   QChronoTimer whose phase drifts against vsync (~2.3us/frame measured), so
   sample->present latency wandered the whole frame interval. Sampling now
   happens in QQuickWindow::afterAnimating (GUI thread, immediately before that
   frame's scene-graph sync), giving a fixed one-frame latency. Measured phase
   wander dropped 2.79px -> 0.10px. afterAnimating rather than frameSwapped
   because it needs no queued hop back from the render thread. The timer stays
   armed as a watchdog and yields while the cadence is alive; that arbitration
   lives in TimelineCadenceArbitrationPolicy.h so it is testable without a
   MainWindow.

2. The scroll was a whole-pixel int. pixelsPerSecond = 120 * zoom puts follow
   playback at ~0.5-4 logical px per frame, so rounding quantised the velocity
   itself: at zoom 0.25 the content froze on half the frames, at 0.75 it
   alternated 1/2px. horizontalScrollValue is now double end to end. Fixing the
   phase alone made this worse, not better - it turned drifting error into white
   error, which a coarse quantiser crosses far more often (autocorrelation
   0.80 -> 0.13, mis-positioned frames 15.9% -> 35.1%). Both changes are needed.

Consumers that genuinely need an integer round at their own boundary: QScrollBar
(int API), the extension-facing state payload (unchanged contract), and the
cull-bucket index, which is now derived through one shared scrollBucketIndex()
because a bare `/` is float division and the two call sites must not disagree.
Scroll targets and the playhead/cursor/entry X use secondToSceneXExact - with a
sub-pixel scroll a rounded playhead no longer cancels the scroll and would
shimmer against the content.

Sub-pixel motion then exposed a second-order problem: unfiltered flat-colour
geometry resamples when translated by a different fraction each frame. The
waveform band and the grid lines therefore snap their translate to the device
pixel grid (not the logical grid - rasterisation happens in device pixels, so a
logical snap costs dpr times more sway for nothing). Note sprites need no such
treatment; QSGTexture::Linear already resamples them correctly.

Measured after both fixes, mis-positioned frames by zoom: 0.25 -> 0.0%,
0.50 -> 0.3-2.1%, 1.50 -> 2.6%, 2.00 -> 4.6%, with mean error a flat ~4.5-5% of
a step at every zoom (proportional sampling jitter, quantisation term gone).

Known residual: notes still use rounded world X while the playhead is exact, so
a note on the playhead can sit up to 0.5 logical px off it. The cursor line and
header markers are flat geometry that is not yet snapped.

Adds scripts/debug/analyze_timeline_cadence.py, which scores a
MIACODE_TIMELINE_HOTPATH_DIAG capture and separates mis-positioned frames from
dropped ones - they are different failures and only the first reads as judder.

+++ src/timeline/TimelineSceneState.h
@@ -113,7 +113,8 @@ struct TimelineSceneState {
     int timelineHeight = 0;
     int laneHeight = 0;
     int laneCount = 0;
-    int horizontalScrollValue = 0;
+    // Sub-pixel — see TimelineSceneBuildRequest::horizontalScrollValue.
+    double horizontalScrollValue = 0.0;
     int headerLeftLimit = 0;
     int headerRightLimit = 0;
     int headerMarkerLeftLimit = 0;
